Intial Version, time difference
This program uses the code/technique in fileage.py to display the age in days since a file was last modified. It allows the program to take a file as argument, it displays the age in days on stdout. It requires an error code on the failure of os.stat